[rename] 임시 파일명 변경
- 파일명 대소문자 변경은 깃이 인식하지 않음
- 이에 임시 파일명으로 변경
